[FIX] 파일 계층 구성 조정
