--- src/main/java/com/wellcom/domain/common/CommonResponse.java
@@ -1,13 +1,0 @@
-package com.wellcom.domain.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import org.springframework.http.HttpStatus;
-
-@Data
-@AllArgsConstructor
-public class CommonResponse {
-    private HttpStatus status;
-    private String message;
-    private Object result;
-}
