+++ src/main/java/com/wellcom/global/auth/login/handler/LoginSuccessHandler.java
@@ -1,0 +1,50 @@
+package com.wellcom.global.auth.login.handler;
+
+import com.wellcom.domain.Member.Repository.MemberRepository;
+import com.wellcom.global.auth.jwt.service.JwtService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Slf4j
+@RequiredArgsConstructor
+public class LoginSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
+
+    private final JwtService jwtService;
+    private final MemberRepository memberRepository;
+
+    @Value("${jwt.access.expiration}")
+    private String accessTokenExpiration;
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
+                                        Authentication authentication) throws IOException {
+        String email = extractUsername(authentication);
+        String role = memberRepository.findByEmail(email).get().getRole().name();
+        String accessToken = jwtService.createAccessToken(email, role);
+        String refreshToken = jwtService.createRefreshToken();
+
+        jwtService.sendAccessAndRefreshToken(response, accessToken, refreshToken);
+
+        memberRepository.findByEmail(email)
+                .ifPresent(user -> {
+                    user.updateRefreshToken(refreshToken);
+                    memberRepository.saveAndFlush(user);
+                });
+        log.info("로그인에 성공하였습니다. 이메일 : {}", email);
+        log.info("로그인에 성공하였습니다. AccessToken : {}", accessToken);
+        log.info("발급된 AccessToken 만료 기간 : {}", accessTokenExpiration);
+    }
+
+    private String extractUsername(Authentication authentication) {
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        return userDetails.getUsername();
+    }
+}
