+++ src/main/java/com/wellcom/global/auth/oauth2/OAuthAttributes.java
@@ -1,0 +1,42 @@
+package com.wellcom.global.auth.oauth2;
+
+import com.wellcom.domain.Member.Member;
+import com.wellcom.domain.Member.Role;
+import com.wellcom.global.auth.oauth2.userinfo.GoogleOAuth2UserInfo;
+import com.wellcom.global.auth.oauth2.userinfo.OAuth2UserInfo;
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.Map;
+import java.util.UUID;
+
+@Getter
+public class OAuthAttributes {
+
+    private String nameAttributeKey; // OAuth2 로그인 진행 시 키가 되는 필드 값, PK와 같은 의미
+    private OAuth2UserInfo oauth2UserInfo; // 소셜 타입별 로그인 유저 정보(닉네임, 이메일, 프로필 사진 등등)
+
+    @Builder
+    public OAuthAttributes(String nameAttributeKey, OAuth2UserInfo oauth2UserInfo) {
+        this.nameAttributeKey = nameAttributeKey;
+        this.oauth2UserInfo = oauth2UserInfo;
+    }
+
+    public static OAuthAttributes ofGoogle(String userNameAttributeName, Map<String, Object> attributes) {
+        return OAuthAttributes.builder()
+                .nameAttributeKey(userNameAttributeName)
+                .oauth2UserInfo(new GoogleOAuth2UserInfo(attributes))
+                .build();
+    }
+
+    public Member toEntity(OAuth2UserInfo oauth2UserInfo) {
+        return Member.builder()
+                .socialType("GOOGLE")
+                .socialId(oauth2UserInfo.getId())
+                .email(UUID.randomUUID() + "@socialUser.com")
+                .nickname(oauth2UserInfo.getNickname())
+                .imageUrl(oauth2UserInfo.getImageUrl())
+                .role(Role.GUEST)
+                .build();
+    }
+}
