+++ src/main/java/com/wellcom/global/auth/oauth2/handler/OAuth2LoginSuccessHandler.java
@@ -1,0 +1,63 @@
+package com.wellcom.global.auth.oauth2.handler;
+
+import com.wellcom.domain.Member.Member;
+import com.wellcom.domain.Member.Role;
+import com.wellcom.domain.Member.Repository.MemberRepository;
+import com.wellcom.global.auth.jwt.service.JwtService;
+import com.wellcom.global.auth.oauth2.CustomOAuth2User;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.stereotype.Component;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+//@Transactional
+public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
+
+    private final JwtService jwtService;
+    private final MemberRepository memberRepository;
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
+        log.info("OAuth2 Login 성공!");
+        try {
+            CustomOAuth2User oAuth2User = (CustomOAuth2User) authentication.getPrincipal();
+
+            if(oAuth2User.getRole() == Role.GUEST) {
+                String accessToken = jwtService.createAccessToken(oAuth2User.getEmail(), Role.GUEST.name());
+                response.addHeader(jwtService.getAccessHeader(), "Bearer " + accessToken);
+                //ToDo: 소셜 로그인 시 회원가입 리다이렉트할 화면 url 설정 필요
+                response.sendRedirect("/sign-up");
+
+                jwtService.sendAccessAndRefreshToken(response, accessToken, null);
+                Member findMember = memberRepository.findByEmail(oAuth2User.getEmail())
+                                .orElseThrow(() -> new IllegalArgumentException("이메일에 해당하는 유저가 없습니다."));
+                findMember.authorizeUser();
+            } else {
+                loginSuccess(response, oAuth2User);
+            }
+        } catch (Exception e) {
+            throw e;
+        }
+
+    }
+
+    // TODO : 소셜 로그인 시에도 무조건 토큰 생성하지 말고 JWT 인증 필터처럼 RefreshToken 유/무에 따라 다르게 처리해보기
+    private void loginSuccess(HttpServletResponse response, CustomOAuth2User oAuth2User) throws IOException {
+        String accessToken = jwtService.createAccessToken(oAuth2User.getEmail(), oAuth2User.getRole().name());
+        String refreshToken = jwtService.createRefreshToken();
+        response.addHeader(jwtService.getAccessHeader(), "Bearer " + accessToken);
+        response.addHeader(jwtService.getRefreshHeader(), "Bearer " + refreshToken);
+
+        jwtService.sendAccessAndRefreshToken(response, accessToken, refreshToken);
+        jwtService.updateRefreshToken(oAuth2User.getEmail(), refreshToken);
+    }
+}
