+++ src/main/java/com/wellcom/global/auth/oauth2/service/CustomOAuth2UserService.java
@@ -1,0 +1,69 @@
+package com.wellcom.global.auth.oauth2.service;
+
+import com.wellcom.domain.Member.Member;
+import com.wellcom.domain.Member.Repository.MemberRepository;
+import com.wellcom.global.auth.oauth2.CustomOAuth2User;
+import com.wellcom.global.auth.oauth2.OAuthAttributes;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
+import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+import org.springframework.stereotype.Service;
+
+import java.util.Collections;
+import java.util.Map;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequest, OAuth2User> {
+
+    private final MemberRepository memberRepository;
+    private final String SOCIAL_TYPE = "GOOGLE";
+
+    @Override
+    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
+        log.info("CustomOAuth2UserService.loadUser() 실행 - OAuth2 로그인 요청 진입");
+
+        OAuth2UserService<OAuth2UserRequest, OAuth2User> delegate = new DefaultOAuth2UserService();
+        OAuth2User oAuth2User = delegate.loadUser(userRequest);
+
+        String registrationId = userRequest.getClientRegistration().getRegistrationId();
+        // OAuth2 로그인 시 PK값
+        String userNameAttributeName = userRequest.getClientRegistration()
+                .getProviderDetails().getUserInfoEndpoint().getUserNameAttributeName();
+        // 소셜 로그인에서 API가 제공하는 userInfo의 Json 값(유저 정보들)
+        Map<String, Object> attributes = oAuth2User.getAttributes();
+
+        OAuthAttributes extractAttributes = OAuthAttributes.ofGoogle(userNameAttributeName, attributes);
+
+        Member createdUser = getUser(extractAttributes);
+
+        return new CustomOAuth2User(
+                Collections.singleton(new SimpleGrantedAuthority(createdUser.getRole().getKey())),
+                attributes,
+                extractAttributes.getNameAttributeKey(),
+                createdUser.getEmail(),
+                createdUser.getRole()
+        );
+    }
+
+    private Member getUser(OAuthAttributes attributes) {
+        Member findUser = memberRepository.findBySocialId(
+                attributes.getOauth2UserInfo().getId()).orElse(null);
+
+        if(findUser == null) {
+            return saveUser(attributes);
+        }
+        return findUser;
+    }
+    
+    private Member saveUser(OAuthAttributes attributes) {
+        Member createdUser = attributes.toEntity(attributes.getOauth2UserInfo());
+        return memberRepository.save(createdUser);
+    }
+}
