+++ src/main/java/com/wellcom/global/auth/oauth2/userinfo/OAuth2UserInfo.java
@@ -1,0 +1,18 @@
+package com.wellcom.global.auth.oauth2.userinfo;
+
+import java.util.Map;
+
+public abstract class OAuth2UserInfo {
+
+    protected Map<String, Object> attributes;
+
+    public OAuth2UserInfo(Map<String, Object> attributes) {
+        this.attributes = attributes;
+    }
+
+    public abstract String getId();
+
+    public abstract String getNickname();
+
+    public abstract String getImageUrl();
+}
