+++ src/main/java/com/wellcom/global/common/CommonResponse.java
@@ -1,0 +1,13 @@
+package com.wellcom.global.common;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import org.springframework.http.HttpStatus;
+
+@Data
+@AllArgsConstructor
+public class CommonResponse {
+    private HttpStatus status;
+    private String message;
+    private Object result;
+}
