+++ src/main/java/com/wellcom/global/common/InitialDataLoader.java
@@ -1,0 +1,32 @@
+package com.wellcom.global.common;
+
+import com.wellcom.domain.Member.Member;
+import com.wellcom.domain.Member.Repository.MemberRepository;
+import com.wellcom.domain.Member.Role;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+
+@Component
+public class InitialDataLoader implements CommandLineRunner {
+    private final MemberRepository memberRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    public InitialDataLoader(MemberRepository memberRepository, PasswordEncoder passwordEncoder) {
+        this.memberRepository = memberRepository;
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    @Override
+    public void run(String... args) throws Exception {
+        if(memberRepository.findByEmail("admin@test.com").isEmpty()){
+            Member adminMember = Member.builder()
+                    .nickname("admin")
+                    .email("admin@test.com")
+                    .password(passwordEncoder.encode("1234"))
+                    .role(Role.ADMIN)
+                    .build();
+            memberRepository.save(adminMember);
+        }
+    }
+}
