+++ src/main/java/com/wellcom/global/config/SecurityConfig.java
@@ -2,15 +2,15 @@ package com.wellcom.global.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.wellcom.domain.Member.Repository.MemberRepository;
-import com.wellcom.global.jwt.filter.JwtAuthenticationProcessingFilter;
-import com.wellcom.global.jwt.service.JwtService;
-import com.wellcom.global.login.filter.CustomJsonUsernamePasswordAuthenticationFilter;
-import com.wellcom.global.login.handler.LoginFailureHandler;
-import com.wellcom.global.login.handler.LoginSuccessHandler;
-import com.wellcom.global.login.service.LoginService;
-import com.wellcom.global.oauth2.handler.OAuth2LoginFailureHandler;
-import com.wellcom.global.oauth2.handler.OAuth2LoginSuccessHandler;
-import com.wellcom.global.oauth2.service.CustomOAuth2UserService;
+import com.wellcom.global.auth.jwt.filter.JwtAuthenticationProcessingFilter;
+import com.wellcom.global.auth.jwt.service.JwtService;
+import com.wellcom.global.auth.login.filter.CustomJsonUsernamePasswordAuthenticationFilter;
+import com.wellcom.global.auth.login.handler.LoginFailureHandler;
+import com.wellcom.global.auth.login.handler.LoginSuccessHandler;
+import com.wellcom.global.auth.login.service.LoginService;
+import com.wellcom.global.auth.oauth2.handler.OAuth2LoginFailureHandler;
+import com.wellcom.global.auth.oauth2.handler.OAuth2LoginSuccessHandler;
+import com.wellcom.global.auth.oauth2.service.CustomOAuth2UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
