--- src/main/java/com/wellcom/global/jwt/filter/JwtAuthenticationProcessingFilter.java
@@ -1,107 +1,0 @@
-package com.wellcom.global.jwt.filter;
-
-import com.wellcom.domain.Member.Repository.MemberRepository;
-import com.wellcom.domain.Member.Member;
-import com.wellcom.global.jwt.service.JwtService;
-import com.wellcom.global.jwt.util.PasswordUtil;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.authority.mapping.GrantedAuthoritiesMapper;
-import org.springframework.security.core.authority.mapping.NullAuthoritiesMapper;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@RequiredArgsConstructor
-@Slf4j
-public class JwtAuthenticationProcessingFilter extends OncePerRequestFilter {
-
-    private static final String[] NO_CHECK_URLS = {"/login"};
-
-    private final JwtService jwtService;
-    private final MemberRepository memberRepository;
-
-    private GrantedAuthoritiesMapper authoritiesMapper = new NullAuthoritiesMapper();
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        for(String url : NO_CHECK_URLS){
-            if (request.getRequestURI().equals(url)) {
-                filterChain.doFilter(request, response);
-                return;
-            }
-        }
-
-        String refreshToken = jwtService.extractRefreshToken(request)
-                .filter(jwtService::isTokenValid)
-                .orElse(null);
-
-        if (refreshToken != null) {
-            checkRefreshTokenAndReIssueAccessToken(response, refreshToken);
-            return;
-        }
-
-        if (refreshToken == null) {
-            checkAccessTokenAndAuthentication(request, response, filterChain);
-        }
-    }
-
-    public void checkRefreshTokenAndReIssueAccessToken(HttpServletResponse response, String refreshToken) {
-        memberRepository.findByRefreshToken(refreshToken)
-                .ifPresent(member -> {
-                    String reIssuedRefreshToken = reIssueRefreshToken(member);
-                    try {
-                        jwtService.sendAccessAndRefreshToken(response, jwtService.createAccessToken(member.getEmail(), member.getRole().name()),
-                                reIssuedRefreshToken);
-                    } catch (IOException e) {
-                        throw new RuntimeException(e);
-                    }
-                });
-    }
-
-    private String reIssueRefreshToken(Member member) {
-        String reIssuedRefreshToken = jwtService.createRefreshToken();
-        member.updateRefreshToken(reIssuedRefreshToken);
-        memberRepository.saveAndFlush(member);
-        return reIssuedRefreshToken;
-    }
-
-    public void checkAccessTokenAndAuthentication(HttpServletRequest request, HttpServletResponse response,
-                                                  FilterChain filterChain) throws ServletException, IOException {
-        log.info("checkAccessTokenAndAuthentication() 호출");
-        jwtService.extractAccessToken(request)
-                .filter(jwtService::isTokenValid)
-                .ifPresent(accessToken -> jwtService.extractEmail(accessToken)
-                        .ifPresent(email -> memberRepository.findByEmail(email)
-                                .ifPresent(this::saveAuthentication)));
-
-        filterChain.doFilter(request, response);
-    }
-
-    public void saveAuthentication(Member curMember) {
-        String password = curMember.getPassword();
-        if (password == null) { // 소셜 로그인 유저의 비밀번호 임의로 설정 하여 소셜 로그인 유저도 인증 되도록 설정
-            password = PasswordUtil.generateRandomPassword();
-        }
-
-        UserDetails userDetailsUser = org.springframework.security.core.userdetails.User.builder()
-                .username(curMember.getEmail())
-                .password(password)
-                .roles(curMember.getRole().name())
-                .build();
-
-        Authentication authentication =
-                new UsernamePasswordAuthenticationToken(userDetailsUser, null,
-                authoritiesMapper.mapAuthorities(userDetailsUser.getAuthorities()));
-
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-    }
-}
